requirements.txt
anthropic>=0.40.0
swebench>=3.0.0
datasets>=2.18.0
docker>=7.0.0
python-dotenv>=1.0.0
